requirements.txt
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart
deepface==0.0.81